[nrf fromtree] drivers: usb: udc: dwc2: split vendor quirks per-vendor
Use one file per vendor quirk. No modification to any content,
only the same code being moved to separate files.

Conflicts in drivers/usb/udc/udc_dwc2_vendor_quirks.h were ignored, i.e.
the file got removed despite having different contents prior to removal.

(cherry picked from commit 57f0805bd5cc9affdd0429a384a9a3d87c0c2276)
